Refactor code, rename import service
