+++ app/controllers/webhooks_controller.rb
@@ -14,7 +14,7 @@ class WebhooksController < ApplicationController
       return render json: {status: 403}, status: :forbidden
     end
 
-    github_webhook_handler(params)
+    WebhookHandler.handle_github(params)
 
     render json: {status: :ok}
   end
@@ -25,7 +25,7 @@ class WebhooksController < ApplicationController
       return render json: {status: 403}, status: :forbidden
     end
 
-    semaphore_webhook_handler(params)
+    WebhookHandler.handle_semaphore(params)
 
     render json: {status: :ok}
   end
@@ -36,48 +36,8 @@ class WebhooksController < ApplicationController
 
   private
 
-  def github_webhook_handler(params)
-    number = NumberExtractor.call(params)
-
-    return unless number.present? && Issue.exists?(id: number)
-
-    PullRequest.auto_create_or_update(params.merge(issue_id: number))
-  end
-
-  def semaphore_webhook_handler(params)
-    range = params[:revision][:branch][:commit_range]
-    branch = params[:revision][:branch][:name]
-    repo = params[:repository][:slug]
-    passed = params[:pipeline][:result] == 'passed'
-    time = params[:pipeline][:done_at]
-
-    first_sha = range.split('...').first
-    last_sha = range.split('...').last
-
-    sha_between = fetch_commit_history(repo, first_sha, last_sha)
-    create_deploys_for_pull_requests(sha_between, branch, passed, time)
-  end
-
   def verify_signature(payload_body, recieved_signature, secret)
     signature = "sha256=#{OpenSSL::HMAC.hexdigest(OpenSSL::Digest.new('sha256'), secret, payload_body)}"
     Rack::Utils.secure_compare(signature, recieved_signature)
-  end
-
-  def fetch_commit_history(repo, first_commit, last_commit)
-    comparison = CLIENT.compare(repo, first_commit, last_commit)
-    comparison.commits.pluck(:sha)
-  end
-
-  def create_deploys_for_pull_requests(sha_between, branch, passed, time)
-    sha_between.each do |sha|
-      pr = PullRequest.find_by(merge_commit_sha: sha)
-      next unless pr
-
-      PullRequestDeployment.auto_create_or_update(branch, pr.id, url, passed, time)
-    end
-  end
-
-  def url
-    "https://#{params[:organization][:name]}.semaphoreci.com/workflows/#{params[:workflow][:id]}/"
   end
 end
