--- test/unit/legacy_import_service_test.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-require_relative '../test_helper'
-class LegacyImportServiceTest < Minitest::Test
-  def test
-    LegacyImportService.new.call
-  end
-end
